backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-dotenv==1.1.*
python-multipart==0.0.*
openai==1.82.*
azure-ai-documentintelligence==1.0.*
azure-cosmos==4.9.*
pydantic==2.11.*
